feat: add UpdateItemViewModel.swift
