+++ SickSangHae/Application/SickSangHaeApp.swift
@@ -11,7 +11,7 @@ import SwiftUI
 struct SickSangHaeApp: App {
     var body: some Scene {
         WindowGroup {
-            UpdateItemView()
+            TabBarView()
         }
     }
 }
